spaceai/spaice-autocl-main/requirements.txt
numpy==1.26.4
pandas==2.2.2
pyyaml==6.0.1
more_itertools==10.2.0
tqdm==4.66.2
torch==2.2.2
ray[tune]== 2.11.0
optuna
scipy==1.13.0
scikit-learn==1.4.2

